apps/users/requirements.txt
Django>=4.1.5<4.2
djangorestframework>=3.14.0<3.15
drf-spectacular>=0.25.1<0.26
psycopg2-binary>=2.9.5<2.10
Faker>=16.6.0<16.7
PyJWT>=2.6.0<2.7